backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
google-generativeai
pydantic
requests
python-docx
PyPDF2
python-multipart
supabase
pycryptodome
jinja2
httpx
markupSafe
twilio
pipecat-ai[websocket,cartesia,google,silero,deepgram,runner,local-smart-turn-v3]>=0.0.105
loguru
pipecatcloud>=0.2.20
